Move package helpers to utils
