+++ packages/core/src/utils/findPackageJsonByName.ts
